scripts/requirements.txt
supabase==1.2.0
python-dotenv==1.0.0 